requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==5.0
django-allauth==0.57.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
